requirements.txt
streamlit
pandas
matplotlib
fuzzywuzzy
python-Levenshtein
requests
pillow
python-dotenv
